fix: device_type接口报错修复 --story=119900119 (merge request !1509)
Squash merge branch 'demand-api-fix' into 'bcc/feature-plan'
fix: device_type接口报错修复 --story=119900119

 



TAPD: --story=119900119

+++ cmd/woa-server/logics/plan/demand.go
@@ -274,7 +274,7 @@ func (c *Controller) UnlockAllResPlanDemand(kt *kit.Kit, crpDemandIDs []int64) e
 func (c *Controller) GetProdResPlanPool(kt *kit.Kit, prodID int64) (ResPlanPool, error) {
 	// get op product all unlocked crp demand ids.
 	opt, err := tools.And(
-		tools.RuleEqual("locked", enumor.CrpDemandUnLocked),
+		tools.RuleEqual("locked", int8(enumor.CrpDemandUnLocked)),
 		tools.RuleEqual("op_product_id", prodID),
 	)
 	if err != nil {
